Added connection events (#1)
This PR adds functionality with connection events using a callback system.

**Problem**
I try to use this library for workshops in which motor equipped devices are controlled (zeppelin, hovercraft, ...). In such cases it is important to have an indication in the device to see if the connection is working or not: motors should be switched off when connection is lost for safety reasons. Also it is interesting to have a LED indicator showing if a device has a connection and is receiving data. Especially in workshops with 10 or 20 devices, it is helpfull when things are not working as expected.

**Solution**
A callback mechanism is added for receiving connection events. Supported events are 
- EmbAjaxConnectionEventConnected: when a browser is connected
- EmbAjaxConnectionEventDisconnected: when no browser is connected anymore
- EmbAjaxConnectionEventMessage: each time some data is received from the browser

The state is connected, as long as at least one device is connected (and a browser session is active). Detection of the connection is done by checking heartbeat (every second).

An example is added with a led connection indicator.